// Front-End/challenge-app/src/app/_shared/models/user.model.ts
export class User {
  id?: string;
  userName?: string;
  email?: string;
  firstName?: string;
  lastName?: string;
  isActive?: boolean;
  isResetPasswordRequired?: boolean;
  initialPassword?: string;
  name?: string;
  userRoles?: string[];
  userGroups?: string[];
  status?: string;
  userRoleDisplay?: string;
  userGroupDisplay?: string;
  password?: string;
}